MauriPay app: pay-by-code, balance, transactions, session auth

+++ lib/feautres/main/data/datasources/main_remote_datasource.dart
@@ -1,3 +1,3 @@
 abstract interface class MainRemoteDatasource {
-  Future<int> getBalance(String user_id);
+  Future<double> getBalance(String user_id);
 }
